% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/lm_sum.R
\name{lm_sum}
\alias{lm_sum}
\title{Cette fonction affiche toutes les statistiques essentielles d'un modèle linéaire classique ou robuste dans un dataframe concis et ordonné}
\usage{
lm_sum(model, plot = F, type = "full")
}
\arguments{
\item{model}{modèle "lm" ou "lm_robust"}

\item{plot}{Par défaut F. F : La sortie est un objet R de type dataframe ; T = la sortie est un plot.}

\item{type}{Par défaut "full". "full" : affiche en sortie les staistiques pour chacun des prédicteurs du modèle et celles du modèle entier ; "pred" : n'affiche que les statistiques des prédicteurs ; "mod" : n'affiche que les statistiques du modèle entier.}
}
\description{
Cette fonction affiche toutes les statistiques essentielles d'un modèle linéaire classique ou robuste dans un dataframe concis et ordonné
}
