% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/lm_check.R
\name{lm_check}
\alias{lm_check}
\title{Cette fonction vérifie les conditions d'application d'un modèle linéaire : normalité, homoscédasticité et indépendance des résidus. De plus, cette fonction réalise une analyse complète des outliers.}
\usage{
lm_check(model)
}
\arguments{
\item{model}{modèle linéaire. Objet R de type "lm" exclusivement.}
}
\description{
Cette fonction vérifie les conditions d'application d'un modèle linéaire : normalité, homoscédasticité et indépendance des résidus. De plus, cette fonction réalise une analyse complète des outliers.
}
