% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/lm_check.R
\name{lm_check}
\alias{lm_check}
\title{This function checks the assumptions for a linear model: normality, homoscedasticity, and independence of the residuals. In addition, this function performs a comprehensive analysis of outliers.}
\usage{
lm_check(model)
}
\arguments{
\item{model}{Linear model. type "lm" only.}
}
\description{
This function checks the assumptions for a linear model: normality, homoscedasticity, and independence of the residuals. In addition, this function performs a comprehensive analysis of outliers.
}
