% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/lm_sum.R
\name{lm_sum}
\alias{lm_sum}
\title{This function displays all the essential statistics for a standard or robust linear model in a concise, well-organized dataframe}
\usage{
lm_sum(model, plot = F, type = "full")
}
\arguments{
\item{model}{Linear model of type "lm" or "lm_robust"}

\item{plot}{Default FALSE. FALSE : output is a dataframe displayed in console ; TRUE = output is a Rplot object.}

\item{type}{Default "full. "full" : displays statistics for each of the model's predictors and for the full model ; "pred" : displays the predictors statistics only ; "mod" : displays the full model statistics only.}
}
\value{
Either a table in your console or a R plot object depending on parameters.
}
\description{
This function displays all the essential statistics for a standard or robust linear model in a concise, well-organized dataframe
}
